v0.4: dependency/CI/tooling hardening + post-audit reliability & secu…

+++ shared/tce_model_gateway/openai_gateway.py
@@ -12,7 +12,7 @@ class OpenAIGateway(ModelGateway):
         api_key: str,
         embed_model: str,
         extract_model: str,
-        timeout: int = 30,
+        timeout: float = 30,
         base_url: str | None = None,
     ) -> None:
         from openai import OpenAI
